requirements.txt
colorama==0.4.6
gitdb==4.0.12
GitPython==3.1.45
smmap==5.0.2
